requirements.txt
openai
supabase
moviepy
pydub
reportlab
llama-index
llama-parse
llama-index-llms-openai
python-dotenv
aiogram>=3.0
fastapi
uvicorn[standard]
python-multipart
aiohttp
toml